+++ image/service/slapd/container-start.sh
@@ -9,6 +9,26 @@ chown -R openldap:openldap /etc/ldap
 # container first start
 if [ ! -e "$FIRST_START_DONE" ]; then
 
+  function get_base_dn(){
+    BASE_DN=""
+    IFS='.' read -ra BASE_DN_TABLE <<< "$LDAP_DOMAIN"
+    for i in "${BASE_DN_TABLE[@]}"; do
+      EXT="dc=$i,"
+      BASE_DN=$BASE_DN$EXT
+    done
+
+    BASE_DN=${BASE_DN::-1}
+  }
+
+  function is_new_schema(){
+    local COUNT=$(ldapsearch -Q -Y EXTERNAL -H ldapi:/// -b cn=schema,cn=config cn | grep -c $1)
+    if [ "$COUNT" -eq 0 ]; then
+      echo 1
+    else
+      echo 0
+    fi
+  }
+
   # database is uninitialized
   if [ -z "$(ls -A /var/lib/ldap)" ]; then
 
@@ -29,11 +49,22 @@ slapd slapd/dump_database select when needed
 EOF
 
     dpkg-reconfigure -f noninteractive slapd
+
+    # start OpenLDAP
+    slapd -h "ldapi:///" -u openldap -g openldap
+
+    get_base_dn 
+    sed -i "s|dc=example,dc=org|$BASE_DN|g" /osixia/slapd/security.ldif
+
+    ldapmodify -Y EXTERNAL -Q -H ldapi:/// -f /osixia/slapd/security.ldif
+
+  else
+
+    # start OpenLDAP
+    slapd -h "ldapi:///" -u openldap -g openldap
+
   fi
 
-
-  # start OpenLDAP
-  slapd -h "ldapi:///" -u openldap -g openldap
 
   # TLS config
   if [ "${USE_TLS,,}" == "true" ]; then
@@ -49,8 +80,7 @@ EOF
     sed -i "s,/osixia/slapd/ssl/ldap.crt,/osixia/slapd/ssl/${SSL_CRT_FILENAME},g" /osixia/slapd/tls.ldif
     sed -i "s,/osixia/slapd/ssl/ldap.key,/osixia/slapd/ssl/${SSL_KEY_FILENAME},g" /osixia/slapd/tls.ldif
 
-    # set tls config
-    ldapmodify -Y EXTERNAL -H ldapi:/// -f /osixia/slapd/tls.ldif -Q
+    ldapmodify -Y EXTERNAL -Q -H ldapi:/// -f /osixia/slapd/tls.ldif
 
     # add localhost route to certificate cn (need docker 1.5.0)
     cn=$(openssl x509 -in /osixia/slapd/ssl/$SSL_CRT_FILENAME -subject -noout | sed -n 's/.*CN=\(.*\)\/*\(.*\)/\1/p')
@@ -60,10 +90,37 @@ EOF
     sed -i "s,TLS_CACERT.*,TLS_CACERT /osixia/slapd/ssl/${SSL_CA_CRT_FILENAME},g" /etc/ldap/ldap.conf
   fi
 
+  # add ppolicy schema if not already exists
+  ADD_PPOLICY=$(is_new_schema ppolicy)
+  if [ "$ADD_PPOLICY" -eq 1 ]; then
+    ldapadd -c -Y EXTERNAL -Q -H ldapi:/// -f /etc/ldap/schema/ppolicy.ldif
+  fi
+
+  # convert  schemas to ldif
+  SCHEMAS=""
+  for f in $(find /osixia/slapd/schema -name \*.schema -type f); do
+    SCHEMAS="$SCHEMAS ${f}"
+  done
+  /osixia/slapd/schema-to-ldif.sh "$SCHEMAS"
+
+  for f in $(find /osixia/slapd/schema -name \*.ldif -type f); do
+    echo "Processing file ${f}"
+    # add schema if not already exists
+    SCHEMA=$(basename "${f}" .ldif)
+    ADD_SCHEMA=$(is_new_schema $SCHEMA)
+    if [ "$ADD_SCHEMA" -eq 1 ]; then
+      echo "add schema ${SCHEMA}"
+      ldapadd -c -Y EXTERNAL -Q -H ldapi:/// -f $f
+    else
+      echo "schema ${f} already exists"
+    fi
+
+  done
+
   # OpenLDAP config 
   for f in $(find /osixia/slapd/config -name \*.ldif -type f); do
-    status "Processing file ${f}"
-    ldapmodify -r -Y EXTERNAL -H ldapi:/// -f $f -Q
+    echo "Processing file ${f}"
+    ldapmodify -Y EXTERNAL -Q -H ldapi:/// -f $f
   done
 
   # stop OpenLDAP