+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+## 0.10.1 (release date: 2015-05-17)
+  - Add ldapi
+  - Add custom ldap schema
+  - Auto convert .schema to .ldif
+  - Fixes issues #2 #3
+
 ## 0.10.0 (release date: 2015-03-03)
   - New version initial release